requirements.txt
python-telegram-bot==20.3
gspread
oauth2client
schedule
python-dotenv
openai
